Moved ProxySettings class to HTTP library module.
Issue: eclipse#423

The class is going to be used to configure the HTTP client.
