+++ core/runtime/src/test/java/org/eclipse/sw360/antenna/util/HttpHelperTest.java
@@ -16,8 +16,7 @@ import org.apache.http.StatusLine;
 import org.apache.http.client.methods.CloseableHttpResponse;
 import org.apache.http.client.methods.HttpGet;
 import org.apache.http.impl.client.CloseableHttpClient;
-import org.eclipse.sw360.antenna.api.configuration.AntennaContext;
-import org.eclipse.sw360.antenna.api.configuration.ToolConfiguration;
+import org.eclipse.sw360.antenna.http.config.ProxySettings;
 import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
